Script updates

+++ scripts/startup.sh
@@ -5,7 +5,13 @@ mkdir -p ${HOME}/jupyter/pentest/base
 rm ${HOME}/jupyter/pentest/base/*.ipynb
 rm -rf ${HOME}/jupyter/pentest/base/tools
 cp ${HOME}/git/nixos/resources/jupyter/pentest/*.ipynb ${HOME}/jupyter/pentest/base
+
 # Mullvad VPN
+if [[ ! -f /etc/vpn-mullvad ]]; then
+  echo "ERROR: Cannot place VPN profiles, could not decrypt?"
+  exit 1
+fi
+
 echo -n "Placing VPN profiles.."
 rm -rf ${HOME}/vpn || true
 /run/current-system/sw/bin/tar xf /etc/vpn-mullvad -C ${HOME} || true
